fix(core): budget arithmetic, line slicing, shared summarizer prompt
