+++ packages/core/src/summarizer/cli.ts
@@ -1,10 +1,11 @@
 import { spawn } from 'node:child_process';
 import type { ContextSlot, ProviderId } from '@goodboy/types';
-import { isSlotKey, SLOT_KEYS, SLOT_LABELS, type SlotKey } from '../context/slots';
+import { isSlotKey, SLOT_KEYS, type SlotKey } from '../context/slots';
 import { computeCostUsd } from '../providers/claude/cost';
 import { PROVIDER_CAPABILITIES } from '../providers/capabilities';
 import { CURSOR_AUTO_MODEL } from '../providers/cursor/models';
 import { inferNextActions, type NextAction, type NextActionsPrState } from './next-actions';
+import { SUMMARIZER_SYSTEM_PROMPT } from './prompt';
 
 type ContextSlotDeltaUpsert = Readonly<{ key: SlotKey; value: string }>;
 
@@ -58,44 +59,6 @@ export class SummarizerParseError extends Error {
     this.name = 'SummarizerParseError';
   }
 }
-
-const SYSTEM_PROMPT = `You maintain a small structured summary for an AI coding session. The summary is the handoff payload, a fresh agent must be able to read these slots alone and continue the work without seeing any prior turns.
-
-There are exactly five slots, each with a stable key:
-${SLOT_KEYS.map((k) => `- ${k} (${SLOT_LABELS[k]})`).join('\n')}
-
-INPUT
-You receive the previous slot values plus the most recent user turn and assistant turn.
-
-CONTEXT PRESERVATION (critical)
-Previous slot content is canonical memory. There is no "append" operation, every upsert REPLACES the slot, so when you change a slot you MUST emit the full merged value (previous content + new additions, with only items the latest turn invalidated removed).
-Never silently drop facts that are still valid. Per slot:
-- goal: refine over time as intent clarifies; do not erase prior framing if still valid.
-- decisions: append new decisions to the existing list. Never remove accepted ones.
-- open_questions: drop only items the latest user turn explicitly resolves. Add new ones only when the assistant is blocked on the user.
-- files_touched: one path per line; append unique paths; never duplicate or drop prior paths unless a file was deleted.
-- last_output_summary: REPLACE, not merge, it summarizes only the latest assistant turn.
-
-FORMATTING (critical, values render as markdown in the UI)
-Each value MUST be compact, well-structured markdown. Never write a wall of prose on one line.
-Rules:
-- Prefer short bullet lists ("- " prefix). One fact per bullet, under ~100 chars; split long bullets in two.
-- Insert a blank line between logically distinct groups of bullets.
-- Use two-space indent + "- " for sub-bullets when nesting context.
-- Use \`backticks\` for identifiers, paths, commands. Use **bold** sparingly for keys/file names that aid scanning.
-- Do NOT use markdown headings (#), the slot label is already the heading.
-- Do NOT wrap the value in code fences unless quoting actual code.
-- Exclude raw tool output and chat-style narration.
-- The "goal" slot stays as one tight sentence or two, no bullets needed unless multiple sub-goals exist.
-
-OUTPUT
-You MUST respond with a single JSON object and nothing else. No prose, no markdown wrapper, no code fences around the JSON.
-The value field is a JSON string; encode newlines inside it as the escape sequence \\n so the rendered markdown breaks across lines.
-Schema:
-{ "upserts": [ { "key": "<one of the five keys>", "value": "<full merged slot value, as compact markdown>" } ] }
-
-Only include slots that actually change. Omit slots that stay the same. Never invent new keys.
-If nothing should change, return { "upserts": [] }.`;
 
 function getCheapModel(providerId: ProviderId): string {
   if (providerId === 'cursor') {
@@ -172,7 +135,7 @@ export class Summarizer {
     userMessage: string,
   ): Promise<{ stdout: string; stderr: string; exitCode: number | null }> {
     return new Promise((resolve, reject) => {
-      const args = buildCliArgs(this.providerId, this.model, SYSTEM_PROMPT, userMessage);
+      const args = buildCliArgs(this.providerId, this.model, SUMMARIZER_SYSTEM_PROMPT, userMessage);
       const child = this.spawnFn(this.binary, args, {
         stdio: ['ignore', 'pipe', 'pipe'],
       });
