+++ packages/core/src/summarizer/client.ts
@@ -1,10 +1,10 @@
 import type { ContextSlot, ProviderId } from '@goodboy/types';
-import { SLOT_BUDGETS } from '../context/budgets';
 import { computeCostUsd } from '../providers/claude/cost';
 import { PROVIDER_CAPABILITIES } from '../providers/capabilities';
 import { CURSOR_AUTO_MODEL } from '../providers/cursor/models';
-import { isSlotKey, SLOT_KEYS, SLOT_LABELS, type SlotKey } from '../context/slots';
+import { isSlotKey, SLOT_KEYS, type SlotKey } from '../context/slots';
 import { inferNextActions, type NextAction, type NextActionsPrState } from './next-actions';
+import { SUMMARIZER_SYSTEM_PROMPT } from './prompt';
 
 export const getCheapModel = (providerId: ProviderId): string => {
   if (providerId === 'cursor') {
@@ -92,49 +92,6 @@ type SummarizeCommandResult = {
   readonly exitCode: number | null;
 };
 
-const SYSTEM_PROMPT = `You maintain a small structured summary for an AI coding session. The summary is the handoff payload, a fresh agent must be able to read these slots alone and continue the work without seeing any prior turns.
-
-There are exactly five slots, each with a stable key:
-${SLOT_KEYS.map((k) => `- ${k} (${SLOT_LABELS[k]})`).join('\n')}
-
-INPUT
-You receive the previous slot values plus the most recent user turn and assistant turn.
-
-CONTEXT PRESERVATION (critical)
-Previous slot content is canonical memory. There is no "append" operation, every upsert REPLACES the slot, so when you change a slot you MUST emit the full merged value (previous content + new additions, with only items the latest turn invalidated removed).
-Never silently drop facts that are still valid. Per slot:
-- goal: the high-level intent of the session, the feature, refactor, or fix being built and why. Keep it stable. Change it only to sharpen or broaden that intent as it clarifies, never to log what just happened. It is NOT a status line. Never add completion markers (done, complete, "review passed", shipped), phase / cluster / step numbers, counts, file paths, or any description of the latest turn's actions. Those belong in last_output_summary or decisions. If the latest turn only reports progress and the underlying intent is unchanged, omit goal from the upserts. If the current goal value already carries status, progress, or per-turn detail, rewrite it down to the clean high-level intent. Never exceed two sentences. If the current value exceeds two sentences, rewrite it down to two sentences or fewer.
-- decisions: append new decisions to the existing list. When a new decision reverses or supersedes an earlier one, REPLACE the earlier entry with the final decision.
-- open_questions: drop only items the latest user turn explicitly resolves. Add new ones only when the assistant is blocked on the user.
-- files_touched: one path per line; append unique paths; never duplicate or drop prior paths unless a file was deleted.
-- last_output_summary: the rolling cumulative TLDR of the whole session. On every turn, REWORK the previous TLDR together with the latest assistant turn into a new TLDR covering what the session has accomplished so far, the current state, and what is in flight. REPLACE is the storage mechanism, so emit the full cumulative TLDR. Do not summarize only the latest turn and do not append a turn-by-turn log.
-
-SLOT BUDGETS (hard maximum characters per emitted value)
-${SLOT_KEYS.map((key) => `- ${key}: ${SLOT_BUDGETS[key]}`).join('\n')}
-
-If a current or updated slot exceeds its budget, emit a compacted full value within the budget. Merge semantic duplicates, replace superseded decisions with the final decision, and keep the most recent and most relevant facts.
-
-FORMATTING (critical, values render as markdown in the UI)
-Each value MUST be compact, well-structured markdown. Never write a wall of prose on one line.
-Rules:
-- Prefer short bullet lists ("- " prefix). One fact per bullet, under ~100 chars; split long bullets in two.
-- Insert a blank line between logically distinct groups of bullets so the UI does not render them squashed.
-- Use two-space indent + "- " for sub-bullets when nesting context.
-- Use \`backticks\` for identifiers, paths, commands. Use **bold** sparingly for keys/file names that aid scanning.
-- Do NOT use markdown headings (#), the slot label is already the heading.
-- Do NOT wrap the value in code fences unless quoting actual code.
-- Exclude raw tool output and chat-style narration.
-- The "goal" slot is one tight high-level sentence (two at most). No bullets, no status, no progress, no per-turn detail.
-
-OUTPUT
-You MUST respond with a single JSON object and nothing else. No prose, no markdown wrapper, no code fences around the JSON.
-The value field is a JSON string; encode newlines inside it as the escape sequence \\n so the rendered markdown breaks across lines.
-Schema:
-{ "upserts": [ { "key": "<one of the five keys>", "value": "<full merged slot value, as compact markdown>" } ] }
-
-Only include slots that actually change. Omit slots that stay the same. Never invent new keys.
-If nothing should change, return { "upserts": [] }.`;
-
 export class Summarizer {
   private readonly providerId: ProviderId;
   private readonly binary: string;
@@ -157,7 +114,7 @@ export class Summarizer {
         model: this.model,
         binary: this.binary,
         userMessage,
-        systemPrompt: SYSTEM_PROMPT,
+        systemPrompt: SUMMARIZER_SYSTEM_PROMPT,
       },
     });
